introduction
During the COVID crisis, we have highlighted practical applications that are contributing to our security - this week, we're going to introduce some machine learning concepts in this pursuit! Face masks have become a mandatory feature in many segments of society and in this workshop we will show you how to create a simple CNN (Convolutional Neural Network) model to automatically detect if someone is wearing one. For this purpose we will use TensorFlow (in combination with Keras). TensorFlow is a free open source library developed by the Google team for data flow and symbolic math. Its main application in practice has been the creation of neural networks. In this session we will practice